common-utils/src/main/java/me/zodac/advent/util/ArrayUtils.java
/*
 * BSD Zero Clause License
 *
 * Copyright (c) 2021-2022 zodac.me
 *
 * Permission to use, copy, modify, and/or distribute this software for any
 * purpose with or without fee is hereby granted.
 *
 * THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
 * WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
 * MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY
 * SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
 * WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
 * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR
 * IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 */

package me.zodac.advent.util;

import java.util.Collection;
import java.util.List;

/**
 * Utility functions for arrays.
 */
public final class ArrayUtils {

    private static final Boolean[] EMPTY_BOOLEAN_ARRAY = new Boolean[0];

    private ArrayUtils() {

    }

    /**
     * Converts the provided {@link Collection} of {@link List}s of {@link Boolean}s to a 2D array.
     *
     * @param input the input {@link Collection} of {@link List}s
     * @return the 2D {@link Boolean} array
     */
    public static Boolean[][] convertToArrayOfArrays(final Collection<? extends List<Boolean>> input) {
        final Boolean[][] array = new Boolean[input.size()][];

        int i = 0;
        for (final List<Boolean> row : input) {
            array[i++] = row.toArray(EMPTY_BOOLEAN_ARRAY);
        }

        return array;
    }

    /**
     * Given an unsorted {@code int} array, finds the smallest index that has a value greater than the {@code thresholdValue}.
     *
     * @param input          the input to check
     * @param thresholdValue the threshold value that any array entry must be greater than
     * @return the smallest index in the {@code int} array
     * @throws IllegalArgumentException thrown if the input is {@code null}, empty or does not have any value greater than the {@code thresholdValue}
     */
    public static int findSmallestIndexGreaterThanThreshold(final int[] input, final int thresholdValue) {
        if (input == null || input.length == 0) {
            throw new IllegalArgumentException("Input cannot be null or empty");
        }

        for (int i = 0; i < input.length; i++) {
            final int houseValue = input[i];

            if (houseValue > thresholdValue) {
                return i;
            }
        }

        throw new IllegalArgumentException(String.format("No value in input is greater than %s", thresholdValue));
    }
}
